oracle/requirements.txt
# Data & ML
numpy>=2.0.0
scipy>=1.14.0
pandas>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0

# API
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.0.0
python-multipart>=0.0.9

# Algorand
py-algorand-sdk>=2.6.0
algokit-utils>=2.0.0

# Puya Smart Contract Compiler
algopy>=0.7.2